+++ src/compute-engine/latex-syntax/dictionary/definitions-complex.ts
@@ -5,16 +5,19 @@ export const DEFINITIONS_COMPLEX: LatexDictionary = [
     name: 'Real',
     kind: 'function',
     latexTrigger: ['\\Re'],
+    arguments: 'implicit',
   },
   {
     name: 'Imaginary',
     kind: 'function',
     latexTrigger: ['\\Im'],
+    arguments: 'implicit',
   },
   {
     name: 'Argument',
     kind: 'function',
     latexTrigger: ['\\arg'],
+    arguments: 'implicit',
   },
   {
     name: 'Conjugate',