+++ src/compute-engine/latex-syntax/dictionary/indexed-types.ts
@@ -63,6 +63,8 @@ export type IndexedFunctionEntry = CommonEntry & {
   kind: 'function';
 
   parse: ExpressionParseHandler;
+
+  arguments?: 'enclosure' | 'implicit';
 };
 
 export type IndexedMatchfixEntry = CommonEntry & {